requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.25.0
PyYAML==6.0.1
rich==13.7.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
asyncio==3.4.3
prometheus-client==0.19.0